% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_individuals.R
\name{plot_individuals}
\alias{plot_individuals}
\title{Plot control graphic Individuals}
\usage{
plot_individuals(x, k, phase, data)
}
\arguments{
\item{x}{Variable of measurements}

\item{k}{Variable for sample index}

\item{phase}{Value to indicate phases range}

\item{data}{Data set with the variables n, d, and k}
}
\value{
A Plotly graphic
}
\description{
Plot control graphic Individuals
}
