% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_u.R
\name{plot_u}
\alias{plot_u}
\title{Plot control graphic U}
\usage{
plot_u(n, c, k, phase, data)
}
\arguments{
\item{n}{Variable for sample size (opportunity area)}

\item{c}{Variable for number of occurrences}

\item{k}{Variable for sample index}

\item{phase}{Value to indicate phases range}

\item{data}{Data set with the variables n, d, and k}
}
\value{
A Plotly graphic
}
\description{
Plot control graphic U
}
