% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xbar.R
\name{plot_xbar}
\alias{plot_xbar}
\title{Plot control graphic X bar}
\usage{
plot_xbar(x, k, data)
}
\arguments{
\item{x}{Variable of measurements}

\item{k}{Variable for sample index}

\item{data}{Data set with the variables n, d, and k}
}
\value{
A Plotly graphic
}
\description{
Plot control graphic X bar
}
