% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_p.R
\name{plot_p}
\alias{plot_p}
\title{Plot control graphic P}
\usage{
plot_p(n, d, k, data)
}
\arguments{
\item{n}{Variable for sample size}

\item{d}{Variable for non-expected observations}

\item{k}{Variable for sample index}

\item{data}{Data set with the variables n, d, and k}
}
\value{
A Plotly graphic
}
\description{
Plot control graphic P
}
