% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_u.R
\name{plot_u}
\alias{plot_u}
\title{Plot control graphic U}
\usage{
plot_u(
  n,
  c,
  k,
  phase,
  append = FALSE,
  index_append = NULL,
  base_plot = NULL,
  data
)
}
\arguments{
\item{n}{Variable for sample size (opportunity area)}

\item{c}{Variable for number of occurrences}

\item{k}{Variable for sample index}

\item{phase}{Value to indicate phases range}

\item{append}{Indicator to generate append at base plot}

\item{base_plot}{Plotly object from base plot}

\item{data}{Data set with the variables n, d, and k}

\item{dt_append}{Value to indicate append phase range}
}
\value{
A Plotly graphic
}
\description{
Plot control graphic U
}
