% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xbar.R
\name{plot_xbar}
\alias{plot_xbar}
\title{Plot control graphic X bar}
\usage{
plot_xbar(
  x,
  k,
  phase,
  append = FALSE,
  index_append = NULL,
  base_plot = NULL,
  data
)
}
\arguments{
\item{x}{Variable of measurements}

\item{k}{Variable for sample index}

\item{phase}{Value to indicate phases range}

\item{append}{Indicator to generate append at base plot}

\item{base_plot}{Plotly object from base plot}

\item{data}{Data set with the variables n, d, and k}

\item{dt_append}{Value to indicate append phase range}
}
\value{
A Plotly graphic
}
\description{
Plot control graphic X bar
}
